% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesspoints.R
\name{accesspoints}
\alias{accesspoints}
\title{accesspoints}
\usage{
accesspoints(plan)
}
\arguments{
\item{plan}{ingressWeb: Ingress web object}
}
\value{
numeric: a vector with accesspoints.
}
\description{
Computes the access points (AP) for a web. Since the AP for building a web depends on current status of the portals, links and fields
\code{accesspoints} can give only very rough estimates. Therefore it is not integrated in \code{summarize}. It will return a vector with
the components
}
\details{
\describe{
\item{\code{portal_ap}}{gives the AP you get if you capture all portals and deploy always eight resonators,}
\item{\code{mod_ap}}{gives the AP if deploy each portal with two mods,}
\item{\code{link_ap}}{gives the AP if you make all links and}
\item{\code{field_ap}}{gives the AP if you make all fields.}
}
}
\examples{
set.seed(0)
web  <- ingressWeb(random2(10))
web2 <- make.links(web, maxarea=10)
accesspoints(web2)
}
