% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycentric.R
\name{barycentric}
\alias{barycentric}
\title{barycentric}
\usage{
barycentric(x, y, xtri, ytri, tolerance = 1e-08)
}
\arguments{
\item{x}{numeric(n): x coordinates}

\item{y}{numeric(n): y coordinates}

\item{xtri}{numeric(3): x coordinates of triangle vertices}

\item{ytri}{numeric(3): y coordinates of triangle vertices}
}
\value{
numeric(n,3): barycentric coordinates for x-y coordinates
}
\description{
cmoputes barycentric coordinates for x-y coordinates related to x-y triangle coordinates
}
\details{
Barycentric are useful for different purposes:
\enumerate{
  \item If all barycentric coordinates are positive then the point is inside the triangle otherwise outside.
  \item If two coordinates are zero andd one coordinate is one then the point is one of the vertices of the triangle.
  \item For an interior point the barycentric coordinates give the proportion of the area if you subtringulate with the trinangle with this point.
}
}
\examples{
x <- random2(15)
barycentric(x[,1], x[,2], x[13:15,1], x[13:15,2])
}
