% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLink.R
\name{checkLink}
\alias{checkLink}
\title{checkLink}
\usage{
checkLink(g, f, t)
}
\arguments{
\item{g}{ingessWeb: ingressWeb}

\item{f}{integer: link source}

\item{t}{integer: link destination}
}
\value{
boolean value
}
\description{
Checks whether in given portal/link structure it is possible to make a link from portal 'f' to portal 't'
}
\examples{
web <- ingressWeb(random2(10)) # empty web
checkLink(web, 1, 2)           # should return TRUE
}
