% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobweb.R
\name{cobweb}
\alias{cobweb}
\title{cobweb}
\usage{
cobweb(n, d = 0)
}
\arguments{
\item{n}{numeric: number of points for one leg}

\item{d}{numeric: distortion of point coordinates (default: 0)}
}
\value{
data frame: with x-y coordinates for cobweb portals
}
\description{
generates positions of \code{3*n} portals for a (distorted) cobweb.
}
\examples{
xy <- cobweb(6)
plot(xy)
#
xy <- cobweb(6, 0.3)
plot(xy)
}
