% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{fields}
\alias{fields}
\title{fields}
\usage{
fields(plan)
}
\arguments{
\item{plan}{ingressWeb: ingress web object}
}
\value{
integer(2): minimum and maximum number of fields
}
\description{
Returns the miminal (no multilayers) and maximal (all possible multilayers) number of the fields of a web.
Depending how you are link you may create or not create multilayers.
}
\examples{
web  <- ingressWeb(random2(10))
fields(web)
web1 <- make.links(web)
fields(web1)
#
web  <- ingressWeb(fishbone(9))
web1 <- make.links(web, fishbone=3)
fields(web1)
}
