% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbone.R
\name{fishbone}
\alias{fishbone}
\title{fishbone}
\usage{
fishbone(n, d = 0)
}
\arguments{
\item{n}{numeric: number of points to generate}

\item{d}{numeric: distortion of point coordinates (default: 0)}
}
\value{
data frame: with x-y coordinates for fish bone web portals
}
\description{
generates positions of \code{n} portals for a (distorted) fishbone web.
}
\examples{
xy <- fishbone(6)
plot(xy)
#
xy <- fishbone(6, 0.5)
plot(xy)
}
