% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingressWeb.R
\name{ingressWeb}
\alias{ingressWeb}
\alias{ingressWeb.default}
\alias{ingressWeb.ingressWeb}
\title{ingressWeb}
\usage{
ingressWeb(x, ...)

\method{ingressWeb}{default}(x, ...)

\method{ingressWeb}{ingressWeb}(x, ...)
}
\arguments{
\item{x}{numeric(n,2): x-y coordinates or ingressWeb object}

\item{...}{further (named) parameters}
}
\value{
a ingressWeb object
}
\description{
creates an empty web from x-y coordinates or updates/an existing web
}
\examples{
web <- ingressWeb(random2(15))
plot(web)
}
