% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{links}
\alias{links}
\title{links}
\usage{
links(plan, perportal = FALSE)
}
\arguments{
\item{plan}{ingressWeb: ingress web object}

\item{perportal}{logical: return links per portal or the whole web (default: FALSE)}
}
\value{
integer: the number of links
}
\description{
Returns the number links for web. For a full web (all portals included) the number should be 3*#portals-6.
}
\examples{
web  <- ingressWeb(random2(10))
links(web)
web1 <- make.links(web)
links(web1)
# Number of links per portal
plot(table(links(web1, TRUE)))
}
