% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.links.R
\name{make.links}
\alias{make.links}
\title{make.links}
\usage{
make.links(plan, ...)
}
\arguments{
\item{plan}{ingressWeb: a ingressWeb with portal positions and no links}

\item{...}{further parameter(s) to generate a linked web}
}
\value{
a linked ingressWeb object
}
\description{
creates a linked web following some linking plan, e.g. cobweb, herring/fish bone, fanfield or maxfield.
Except for the maxfield (default) you need to provide further parameters, see Details.
}
\details{
In the Ingress community exists a number of popular linking plans. The \href{https://www.youtube.com/watch?v=FkHtelZxgtg}{videos of Michael Hartley} are very helpful
for the herring bone or fish bone, the cobweb, the fanfield and for the \href{http://www.ingress-maxfield.com/}{maxfield}.
To realize these linking plans you neet to set further named parameters (see also \code{vignette("ingressWeb")}):
\describe{
  \item{cobweb}{provide either \code{cobweb=1}, \code{cobweb=2} or \code{cobweb=3} depending at which corner the cobweb should start.}
  \item{fishbone}{provide either \code{fishbone=1}, \code{fishbone=2} or \code{fishbone=3} depending at which corner the fish bone web should start.}
  \item{fanfield}{provide either \code{fanfield=1}, \code{fanfield=-1}, \code{fanfield=2}, \code{fanfield=-2}, \code{fanfield=3}, \code{fanfield=-3} depending at which corner the fish bone web should start. A negavtive value reverses the order.}
  \item{delaunay}{provide \code{delaunay=TRUE} for a \href{https://en.wikipedia.org/wiki/Delaunay_triangulation}{Delaunay} web. This web contains usually one multilayer field.}
}
}
\examples{
web <- ingressWeb(random2(15))
web <- make.links(web)
plot(web, main="Maxarea web")
# Fish/herring bone
web <- ingressWeb(fishbone(8))
web1 <- make.links(web, fishbone=1)
plot(web1, main="Fishbone 1 web")
web2 <- make.links(web, fishbone=2)
plot(web2, main="Fishbone 2 web")
web3 <- make.links(web, fishbone=3)
plot(web3, main="Fishbone 3 web")
# Cobweb
web <- ingressWeb(cobweb(6))
web1 <- make.links(web, cobweb=1)
plot(web1, main="Cobweb 1")
web2 <- make.links(web, cobweb=2)
plot(web2, main="Cobweb 2")
web3 <- make.links(web, cobweb=3)
plot(web3, main="Cobweb 3")
# Fanfield
web <- ingressWeb(random2(10))
web1 <- make.links(web, fanfield=1)
plot(web1, main="Fanfield 1")
web2 <- make.links(web, fanfield=-1)
plot(web2, main="Fanfield -1")
web3 <- make.links(web, fanfield=2)
plot(web3, main="Fanfield 2")
# Delaunay triangulation
web <- ingressWeb(random2(10))
web1 <- make.links(web, delaunay=TRUE)
}
