% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindunits.R
\name{mindunits}
\alias{mindunits}
\title{mindunits}
\usage{
mindunits(plan, density = 1)
}
\arguments{
\item{plan}{ingressWeb: ingress web object}

\item{density}{numeric: mutliplicator for the mindunit density (default: 1)}
}
\value{
numeric(2): Mimimal and maximal area of all fields (if density=1)
}
\description{
Returns the miminal (no multilayers) and maximal (all possible multilayers) mindunits of a web.
The mindunits in Ingress are calculated by the area of the fields multiplied with the mindunits density.
The mindunits density depends on the population living in that area. Exact values for the mindunit density
are not known for the public. If density is set to one (default) then area of the fields will calculated.
}
\examples{
web  <- ingressWeb(random2(10))
mindunits(web)
web1 <- make.links(web)
mindunits(web1)
}
