% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ingressWeb.R
\name{plot.ingressWeb}
\alias{plot.ingressWeb}
\title{plot}
\usage{
\method{plot}{ingressWeb}(x, text = FALSE, ...)
}
\arguments{
\item{x}{ingressWeb: the web}

\item{text}{logical: should the portal number plotted (default: FALSE)}

\item{...}{further parameter for plotting the portal position (default: asp=TRUE, xlab="x", ylab="y", pch=19, axes=FALSE)}
}
\value{
a plot with the linked web
}
\description{
plots a web
}
\examples{
web <- ingressWeb(random2(15))
plot(web, main="Unlinked web")
}
