% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarIndex.R
\name{polarIndex}
\alias{polarIndex}
\title{polarIndex}
\usage{
polarIndex(x, y, i)
}
\arguments{
\item{x}{numeric(n): x coordinates}

\item{y}{numeric(n): y coordinates}

\item{i}{numeric: index of observation}
}
\value{
index or ordered points
}
\description{
Under the assumption that i is the index of the outer triangle of the x-y coordinates, it computes the order of the datapoint
with decreasing/increasing order by the (polar) angles in relation to point i. A negative value of i reverses the order
}
\examples{
x <- random2(15)
plot(x, asp=TRUE)
polarIndex(x[,1], x[,2], 15)
polarIndex(x[,1], x[,2], -15)
}
