% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portals.R
\name{portals}
\alias{portals}
\title{portals}
\usage{
portals(plan)
}
\arguments{
\item{plan}{ingressWeb: ingress web object}
}
\value{
integer: the number of portals
}
\description{
Returns the number portals for a web.
}
\examples{
web  <- ingressWeb(random2(10))
portals(web)
}
