% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random2.R
\name{random2}
\alias{random2}
\title{random2}
\usage{
random2(n, d = 1.05, random = "rnorm", ...)
}
\arguments{
\item{n}{numeric: number of points for one leg}

\item{d}{numeric: distance for outer triangle (default: 1.05)}

\item{random}{function: univariate random number generator (default: \code{\link{rnorm}})}

\item{...}{further parameters for \code{random}}
}
\value{
x-y coordinates portals for randomly generated web.
}
\description{
generates random positions of \code{n} portals for a web.
}
\examples{
xy <- random2(6)
plot(xy)
#
xy <- random2(6, random="runif")
plot(xy)
}
