% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyWeb.R
\name{shinyWeb}
\alias{shinyWeb}
\title{shinyWeb}
\usage{
shinyWeb()
}
\value{
runs the Shiny app
}
\description{
An interactive Shiny app to create portal positions and create webs.
}
\details{
Note to run \code{shinyWeb} you should install the packages \code{shiny} and \code{shinydashboard} before.
}
\examples{
\dontrun{
shinyWeb()
}
}
