% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ingressWeb.R
\name{summary.ingressWeb}
\alias{summary.ingressWeb}
\title{summary.ingressWeb}
\usage{
\method{summary}{ingressWeb}(object, ...)
}
\arguments{
\item{object}{ingressWeb: ingress web o}

\item{...}{further parameters}
}
\value{
numeric: summary data about the web
}
\description{
Computes a summary for a web. Note that the mindunits are rounded up to the next integer number.
}
\examples{

web  <- ingressWeb(fishbone(9))
summary(web)
web1 <- make.links(web, fishbone=3)
summary(web1)
}
