% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangulate.R
\name{triangulate}
\alias{triangulate}
\title{triangulate}
\usage{
triangulate(plan, base, portals = NULL)
}
\arguments{
\item{plan}{ingressWeb: an ingers web object}

\item{base}{numeric: the number of the base portal for the triangulation}

\item{portals}{numeric:}
}
\value{
a set or triangles
}
\description{
Creates a tringulation of the convex hull
}
\examples{
# Read portal position from a Excel CSV file
path     <- system.file('shiny', 'data', package='ingressWeb')
filename <- paste(path, 'Leisepark-170212.csv', sep='/')
portals  <- read.csv(filename)
# convert latitude and longitude to x-y coordinates using the
# Universal Transverse Mercator coordinate system
xy     <- ll2xy(portals, utm=33)
web    <- ingressWeb(xy)
webtri <- make.links(web, triangle=33)
plot(webtri)
}
